# openapi.yaml
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/check":{"post":{"tags":["Отвечает за обработку запросов, связанных с офисам/отделениями"],"summary":"Выдача оптимального отделения","description":"Получает список задач клиента\n    Отдает оптимальное отделение","operationId":"getMeOptimalOffice","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskListRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OptimalOfficeResponse"}}}}}}},"/api/all":{"get":{"tags":["Отвечает за обработку запросов, связанных с офисам/отделениями"],"operationId":"getMeOptimalOffice_1","parameters":[{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/OfficesResponse"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OfficesResponse"}}}}}}}},"components":{"schemas":{"TaskListRequest":{"type":"object","properties":{"pointCoordinates":{"type":"string","description":"координаты клиента в формате 60.497874,56.926760 "},"cardIssue":{"type":"boolean","description":"true/false выпуск карт "},"withdrawCash":{"type":"boolean","description":"true/false снятие наличных "},"currencyExchange":{"type":"boolean","description":"true/false обмен валют"}},"description":"Входящий запрос с координатами клиента и со списком задач"},"OptimalOfficeResponse":{"type":"object","properties":{"uri":{"type":"string","description":"координаты отделения"},"workload":{"type":"integer","description":"загруженность отделения","format":"int32"}},"description":"Ответ API включает себя координаты оптимального отделения и загрузку этого отделения"},"Office":{"type":"object","properties":{"coordinates":{"type":"array","description":"координаты офиса в виде = [60.523104,56.869266]","items":{"type":"string","description":"координаты офиса в виде = [60.523104,56.869266]"}},"workload":{"type":"integer","description":"Загрузка офиса в виде = 99","format":"int32"}},"description":"координаты офиса и его загрузка"},"OfficesResponse":{"type":"object","properties":{"offices":{"type":"array","description":"Список координат отделений с загрузкой","items":{"$ref":"#/components/schemas/Office"}}},"description":"Координаты отделений с загрузкой"}}}}